Remove moot conversion
I'd originally thought to make the one operation Rotate(i32), with the
sign indicating right or left.  Then on seeing we have
std::slice.rotate_right() and rotate_left I split the ops into the
RotateRight(usize) and Left we have now, but the 'as usize' has stuck
around from the previous edit.

Perhaps odd that clippy didn't call this out, spotted it reading back
over the code.

+++ src/day21.rs
@@ -65,8 +65,8 @@ impl Scrambler {
                     *c = x as u8
                 }
             }),
-            Instr::RotateLeft(distance) => bytes.rotate_left(distance as usize),
-            Instr::RotateRight(distance) => bytes.rotate_right(distance as usize),
+            Instr::RotateLeft(distance) => bytes.rotate_left(distance),
+            Instr::RotateRight(distance) => bytes.rotate_right(distance),
             Instr::RotateByLetter(c) => {
                 if let Some(index) = bytes.iter().position(|b| *b == c as u8) {
                     let distance = if index >= 4 { index + 2 } else { index + 1 };
